replaced transition and derive macro
